% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{plot_multi_images}
\alias{plot_multi_images}
\title{Plot multiple images}
\usage{
plot_multi_images(list_images, par_ROWS, par_COLS, ...)
}
\arguments{
\item{list_images}{a list of images that should be visualized}

\item{par_ROWS}{an integer specifying the number of rows of the ouput plot-grid}

\item{par_COLS}{an integer specifying the number of columns of the output plot-grid}

\item{...}{further arguments for the 'plot_multi_images' method of the 'GaborFeatureExtract' R6 class ('OpenImageR' package)}
}
\value{
it doesn't return an R object but it displays a list of input images
}
\description{
Plot multiple images
}
\details{
For the usage of the 'plot_multi_images()' function see the example section of the 'fastGLCM_Rcpp()' function and 'fastglcm()' R6 class
}
